my brain is fried like a chicken nugget. pick this up again tomorrow
