--- fixtures/harperdb-base-component/package.json
@@ -1,5 +1,0 @@
-{
-	"name": "harperdb-base-component",
-	"private": true,
-	"type": "module"
-}
