--- fixtures/harperdb-base-component/schema.graphql
@@ -1,5 +1,0 @@
-type Dog @table @export {
-	id: ID @primaryKey
-	name: String
-	breed: String
-}
