+++ fixtures/next-13/config.yaml
@@ -1,3 +1,2 @@
-'@harperdb/nextjs':
-  package: '@harperdb/nextjs'
-  files: '/*'
+'@harperfast/nextjs':
+  package: '@harperfast/nextjs'
