--- fixtures/next-9/.node-version
@@ -1,1 +1,0 @@
-16.20.2
