--- fixtures/next-9/check-node-version.js
@@ -1,7 +1,0 @@
-const required = '16';
-const current = process.version.slice(1).split('.')[0];
-
-if (current !== required) {
-	console.error(`Error: Node version ${required} is required, but you're running version ${current}`);
-	process.exit(1);
-}
