--- fixtures/next-9/config.yaml
@@ -1,3 +1,0 @@
-'@harperdb/nextjs':
-  package: '@harperdb/nextjs'
-  files: '/*'
