--- fixtures/next-9/next.config.js
@@ -1,1 +1,0 @@
-module.exports = {};
