--- fixtures/next-9/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "next-9",
-	"private": true,
-	"scripts": {
-		"dev": "next dev",
-		"build": "next build",
-		"start": "next start",
-		"lint": "next lint",
-		"postinstall": "npm link harperdb && npx playwright install chromium",
-		"pretest": "node check-node-version.js",
-		"test": "playwright test"
-	},
-	"dependencies": {
-		"@harperdb/nextjs": "file:../../",
-		"next": "9.5.5",
-		"react": "16.14.0",
-		"react-dom": "16.14.0"
-	},
-	"devDependencies": {
-		"@playwright/test": "1.49.1",
-		"test-utils": "file:../../test-utils/"
-	},
-	"engines": {
-		"node": "16.x"
-	}
-}
