--- fixtures/next-9/pages/_document.js
@@ -1,15 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
-
-export default class MyDocument extends Document {
-	render() {
-		return (
-			<Html>
-				<Head />
-				<body>
-					<Main />
-					<NextScript />
-				</body>
-			</Html>
-		);
-	}
-}
