--- fixtures/next-9/pages/page-2.js
@@ -1,10 +1,0 @@
-import Link from 'next/link';
-
-export default function Page2() {
-	return (
-		<div>
-			<h1>Page 2</h1>
-			<Link href="/">Home</Link>
-		</div>
-	);
-}
