+++ package.json
@@ -1,10 +1,10 @@
 {
 	"name": "@harperfast/nextjs",
-	"version": "1.2.0",
+	"version": "2.0.0",
 	"type": "module",
 	"description": "A Harper plugin for running Next.js apps.",
 	"keywords": [
-		"harperdb",
+		"harper",
 		"next.js",
 		"distributed",
 		"applications"
@@ -12,11 +12,11 @@
 	"homepage": "https://github.com/HarperFast/nextjs",
 	"bugs": {
 		"url": "https://github.com/HarperFast/nextjs/issues",
-		"email": "support@harperdb.io"
+		"email": "opensource@harperdb.io"
 	},
 	"author": {
-		"name": "HarperDB",
-		"email": "support@harperdb.io"
+		"name": "Harper",
+		"email": "opensource@harperdb.io"
 	},
 	"repository": {
 		"type": "git",
@@ -24,8 +24,8 @@
 	},
 	"license": "Apache-2.0",
 	"files": [
-		"config.yaml",
 		"dist",
+		"config.yaml",
 		"schema.graphql"
 	],
 	"scripts": {
@@ -33,7 +33,7 @@
 		"format:check": "npm run format -- --check",
 		"format:fix": "npm run format -- --write",
 		"lint": "eslint .",
-		"build": "tsc",
+		"build": "tsc -p tsconfig.build.json",
 		"test:integration": "playwright test --config integrationTests/playwright.config.ts"
 	},
 	"engines": {
@@ -53,7 +53,7 @@
 		"@types/node": "22.19.11",
 		"eslint": "9.17.0",
 		"globals": "15.14.0",
-		"harper": "file:../harper",
+		"harper": "5.0.0",
 		"next": "15.5.12",
 		"prettier": "3.4.2",
 		"typescript": "5.9.3"
